+++ src/ion/modules/map/functions/map_hash.h
@@ -1,4 +1,4 @@
-static inline u64 map_hash (
+static inline u32 map_hash (
     void* key,
-    u64 key_typesize
+    u32 key_typesize
 );
